// src/types/intelligent.ts
export enum Intelligent {
  WORD = 'word',
  LOGIC = 'logic',
  SPACE = 'space',
  MOVE = 'move',
  MUSIC = 'music',
  INNER = 'inner',
  COMMUNICATION = 'communication',
}

// export enum Intelligent {
//   WORD = 'WORD',
//   LOGIC = 'LOGIC',
//   SPACE = 'SPACE',
//   MOVE = 'MOVE',
//   MUSIC = 'MUSIC',
//   INNER = 'INNER',
//   COMMUNICATION = 'COMMUNICATION',
// }